Centralize code add placeholders

+++ includes/save_file.php
@@ -1,5 +1,5 @@
 <?php
-define('ALLSKY_HOME', '/home/pi/allsky');                     // value updated during installation
+define('ALLSKY_HOME', 'XX_ALLSKY_HOME_XX');		// value updated during installation
 
 $content = "";
 $path = "";
@@ -19,14 +19,20 @@ if ($path == "") {
 }
 
 // "current" is a web alias to ALLSKY_HOME.
+// $path is the web address to the file; we need the physical path ($file) to move.
 $file = str_replace('current', ALLSKY_HOME, $path);
+if (! file_exists($file)) {
+	echo "save_file.php: file to save '$file' does not exist!";
+	exit;
+}
+
 $tempFile = getcwd() . "/temp";
 if (file_put_contents($tempFile, $content) == false) {
 	echo error_get_last()['message'];
 	exit;
 } else {
 	// shell_exec() doesn't return anything unless there is an error.
-	$msg = shell_exec("sudo mv '$tempFile' '$file' || echo 'Unable to mv [$tempFile] to [$file]'");
+	$msg = shell_exec("x=\$(sudo mv '$tempFile' '$file' 2>&1) || echo 'Unable to mv [$tempFile] to [$file]': \${x}");
 	if ($msg == "") {
 		shell_exec("sudo chown pi:pi '$file'; sudo chmod +x '$file'");
 	} else {
